Tested system by mocking up middleware with service
Added a provided service to the Smart Product for convenience. This way
the library will automatically register a system instance in the
Arrowhead database and we do not have to do that step manually. Also it
can offer usefull services as an echo, and a test service that has been
used to run test without the need to setup a Middleware system.

The command line interface has been expanded with an exit option, to
shutdown the system gracefully.

The main class code has been polished after testing, and it seems to
work.